Refactor: 코드래빗 피드백 반영
