+++ common/src/main/java/com/sparta/moim/common/passport/Passport.java
@@ -1,4 +1,4 @@
-package com.sparta.moim.common.passport.enums;
+package com.sparta.moim.common.passport;
 
 public record Passport(
     String userTrackingId,
